requirements.txt
aniso8601==10.0.0
bcrypt==4.3.0
bitarray==3.1.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.0
Flask-RESTful==0.3.10
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mmh3==5.1.0
numpy==2.2.3
pandas==2.2.3
paramiko==3.5.1
pycparser==2.22
PyNaCl==1.5.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
